completed charge calculation

+++ api/src/main/java/com/toolrentalstore/toolrentalpos/ToolRentalPos/Controllers/ShopOrderController.java
@@ -83,7 +83,7 @@ public class ShopOrderController implements BaseController {
 
     @PostMapping(value = "/orders/calc-cart-price") 
     @ResponseBody
-    public Float calculateCartTotal(@RequestBody Cart cart ) {   
+    public double calculateCartTotal(@RequestBody Cart cart ) {   
 
         return orderService.calculateTotalForCart(cart);
     }
